requirements.txt
websocket-client
threaded
requests
emoji